Move key handler for scroll view into a dedicated class
This will allow the scrolling functionality to also be used in subclasses of UIKit’s scroll view subclasses.

Functionality that can be has been moved into an extension on UIScrollView that the new key handler uses.

One behaviour change here is that the custom scrolling animation is only used for keyboard-driven scrolling. It no longer overrides setContentOffset:animated: without calling super. This means keyboard scrolling with another animated scrolling won’t work as well but generally better to not mess with stuff like is was before.