Improve localisations (#7)
This improves the localisations added in https://github.com/douglashill/KeyboardKit/pull/6. This was done by looking at:

- the previously generated text
- the bar button item accessibility labels
- the most common translations across Apple’s strings files
- running iOS in many languages and seeing what text is used
- dictionaries

There is no clear superior source in the Apple translations, but combining the sources of information above with judgment should have resulted in better text.

This branch consolidates each of Delete and Refresh to a single entry each since the intended meaning and context is the same.

Lastly, the ExtractLocalisedStrings target has been removed because the strings files need heavy manual editing. The script is a tool to assist rather than full automation so including it in this repository does not make sense.