Fix collection views trying to select non-selectable elements
Fixes arrow key navigation in lists with UICollectionView not working because after the third cell it would ‘select’ the background decoration view which has index path (0, 0) so the selection jumped back to the first cell.

The actual fix for that is checking the `representedElementCategory`. Also fixes the delegate not being respected, and checks alpha just in case because the user might be confused if an invisible item is selected.