Move collapsing to focused column into framework
And remove public API for getting focused view controller to pare it down to the minimal public API.

And remove expanding to a display mode that shows the view that was displayed when collapsed. While I think this is nice, it’s not related to the focused column at all. It’s something I think UISplitViewController should do, but that does not mean split views in KeyboardKit should do it.