Work around tests not being able to run on macOS 11.0.1
Using Xcode 12.3 on macOS 11.0.1, the tests would not run because Xcode seemed to be checking the macOS deployment target when it should be checking the iOS deployment target. This was only for tests: running the demo app directly worked fine.

Work around this Catalyst bug by specifying some older macOS deployment target.