Use simpler solution for scrolling limits
This reverts to be more similar to before 3980fb748fb71274d2f5f586477653611af4b005.

That commit fixed handling when the content was smaller than the bounds, but it regressed in how it handled content inset.

What was there before was nearly correct, it just needed to apply the minimum last instead of first.